==> 00_Notebooks/02_scraping_web.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import requests

url="https://en.wikipedia.org/wiki/Community_areas_in_Chicago"
headers = {"User-Agent": "Mozilla/5.0(compatible;MyScraper/1.0)"}

response = requests.get(url=url, headers=headers)

soup = BeautifulSoup(response.content, 'html.parser')

# print(soup)

title = soup.find(id='firstHeading')

for table in soup.find_all("table"):
    print(table.get("class"))

table = soup.find("table", class_ = "wikitable sortable plainworheaders mwd-atatable")

for row in table.tbody.findAll[2:-1]:
    cells = row.find_all("tr")
    print(row)
    print("-"*100)

for row in table.tbody.findAll[2:-1]:
    header_cell = row.find("th")
    print(header_cell.a.get_text(Strip=True))

data = []

for row in table.tbody.findAll[2:-1]:
    cells = row.find_all("tr")
    header_cell = row.find("th")
    area_code = cells[0].get_text(strip=True)
    community_name = header_cell.a.get_text(Strip=True)

    data.append({"area_code": area_code, "community_name": community_name})

community_areas = pd.DataFrame(data)

community_areas["area_code"] = community_areas["area_code"].astype(int)

community_areas.to_csv("community_areas_master.csv", index=False)
